Post Controller Update: Check If User Liked Post

+++ InstagramClone/API/PostService.swift
@@ -59,12 +59,21 @@ struct  PostService {
     
     static func unlikePost(post: Post, completion: @escaping(FirestoreCompletion)){
         guard let uid = Auth.auth().currentUser?.uid else { return }
-       // guard post.likes > 0 else { return }
+        guard post.likes > 0 else { return }
         
         COLLETION_POSTS.document(post.postId).updateData(["likes": post.likes - 1])
         
         COLLETION_POSTS.document(post.postId).collection("post-likes").document(uid).delete() {_ in
             COLLECTION_USERS.document(uid).collection("user-likes").document(post.postId).delete(completion: completion)
+        }
+    }
+    
+    static func checkIfUserLikedPost(post: Post, completion: @escaping(Bool) -> Void){
+        guard let uid = Auth.auth().currentUser?.uid else { return }
+
+        COLLECTION_USERS.document(uid).collection("user-likes").document(post.postId).getDocument { (snapshot, _) in
+            guard let didLike = snapshot?.exists else { return }
+            completion(didLike)
         }
     }
